Changes app name and apply review changes
