+++ geonode/management_commands_http/management/commands/__init__.py
@@ -16,12 +16,3 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 #########################################################################
-from django.apps import AppConfig
-
-
-class MngmtCommandsHttpAppConfig(AppConfig):
-    name = 'geonode.mngmt_commands_http'
-    verbose_name = "Management Commands Over HTTP"
-
-
-default_app_config = 'geonode.mngmt_commands_http.MngmtCommandsHttpAppConfig'
