+++ geonode/urls.py
@@ -171,7 +171,7 @@ urlpatterns += [
     url(r'^api/users', users, name='users'),
     url(r'^api/v2/', include(router.urls)),
     url(r'^api/v2/', include('geonode.api.urls')),
-    url(r'^api/v2/', include('geonode.mngmt_commands_http.urls')),
+    url(r'^api/v2/', include('geonode.management_commands_http.urls')),
     url(r'^api/v2/api-auth/', include('rest_framework.urls', namespace='geonode_rest_framework')),
     url(r'', include(api.urls)),
